fix integration slider
